Format string safety fixes, test case for NSData<->D-Bus byte arrays conversions (and a accompanying bugfix).
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/dbuskit/trunk@37559 72102866-910b-0410-8b05-ffd578937521

+++ Source/DKMessage.m
@@ -95,7 +95,7 @@
   if (res != NULL)
   {
     NSDebugMLog(@"Sending preallocated message (type: %ld, signature, '%s') via endpoint %@",
-      dbus_message_get_type(msg),
+      (long)dbus_message_get_type(msg),
       dbus_message_get_signature(msg),
       endpoint);
     // We have preallocated resources and use those to send the message
